% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goodbye.R
\name{goodbye}
\alias{goodbye}
\title{Goodbye, world!}
\usage{
goodbye()
}
\arguments{
\item{none}{required}
}
\value{
'Goodbye, world!'.
}
\description{
This is an example function named 'goodbye'
which prints 'Goodbye, world!'.
}
\examples{
goodbye()
}
