% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safe_names.R
\name{safe_names}
\alias{safe_names}
\title{Make database friendly names}
\usage{
safe_names(names)
}
\arguments{
\item{names}{a vector of column names}
}
\value{

}
\description{
janitor::clean_names is probably better to use. But ive been using this for
ages.
}
\examples{
\dontrun{
safe_names(c("Bad name!", "And&an0ther@!"))
}
}
